# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Detail": {
        "description": "FastAPI's error body, written down so it can be named in `responses=`.\n\nEvery non-validation error this API returns is one of these. Declaring it\nis what turns \"401\" in the docs from a bare number into a shape a client\ncan code against.",
        "example": {
          "detail": "Could not validate credentials"
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Detail",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PostOut": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "published": {
            "default": true,
            "title": "Published",
            "type": "boolean"
          },
          "saved": {
            "default": false,
            "title": "Saved",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "voted": {
            "default": false,
            "title": "Voted",
            "type": "boolean"
          },
          "votes": {
            "default": 0,
            "title": "Votes",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "content",
          "id",
          "created_at",
          "updated_at",
          "user_id",
          "user"
        ],
        "title": "PostOut",
        "type": "object"
      },
      "PostPage": {
        "description": "A page of posts plus the metadata a UI needs to render pagination.",
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "has_prev": {
            "title": "Has Prev",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages",
          "has_next",
          "has_prev"
        ],
        "title": "PostPage",
        "type": "object"
      },
...
      "UserOut": {
        "description": "A person, as everybody else sees them. No email: that's a credential,\nand handing it to anyone who can walk a list of ids is how address books\nget scraped.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Source on GitHub",
      "url": "https://github.com/AndrewTechTips/social-feed-application"
    },
    "description": "The backend behind **Commons** — a small public feed.\n\nPosts, comments, upvotes and accounts, with a two-token auth flow. Reading is public; writing needs an account.\n\nThe published demo at [andrewtechtips.github.io/social-feed-application](https://andrewtechtips.github.io/social-feed-application/) reimplements this same contract in the browser, because a static host has nowhere to run FastAPI. This document is the contract both of them answer to.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/AndrewTechTips/social-feed-application/blob/main/LICENSE"
    },
    "title": "Social Feed API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/posts/{id}/save": {
      "delete": {
        "description": "Remove a post from your shelf.\n\nNo 404 on this one, deliberately, and not for symmetry: there is nothing\nhere worth reporting as missing. Whether the post exists, whether you could\nsee it, and whether you had saved it all end in the same place — it is not\non your shelf — and three ways of saying that would be three branches a\nclient has to handle to reach one outcome.",
        "operationId": "unsave_post_api_v1_posts__id__save_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Off your shelf — including when it was never on it. The state you asked for is the state you're in."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Take a post off your shelf",
        "tags": [
          "Shelf"
        ]
      },
      "put": {
        "description": "Save a post you can see.\n\nA draft you don't own answers 404, exactly as reading it would — a 403\nwould confirm that it exists.",
        "operationId": "save_post_api_v1_posts__id__save_put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "On your shelf. Saying so again is the same answer — this is a set, and asking for a post to be in it twice is not an error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No such row — or one you aren't allowed to know about. Somebody else's unpublished post answers this, deliberately, because a 403 would confirm that it exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Put a post on your shelf",
        "tags": [
          "Shelf"
        ]
      }
    },
...
    "/api/v1/shelf": {
      "get": {
        "description": "Your shelf, most recently saved first.\n\nOrdered by when you saved it rather than by when it was written, which is\nthe difference between a shelf and a feed: what you put there last is what\nyou meant to read next. `ix_saves_user_id_created_at` hands that ordering\nback without a sort.\n\nIt is one page of the feed's own machinery, so the visibility rule, the\nvote counts and `voted` are all the same here as anywhere else.",
        "operationId": "get_shelf_api_v1_shelf_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "has_next": false,
                  "has_prev": false,
                  "items": [
                    {
                      "content": "There's a reading room near the old tram depot that never closes. No membership, no desk, no one checking whether you belong there.",
                      "created_at": "2026-09-08T22:41:07.220Z",
                      "id": 12,
                      "published": true,
                      "saved": true,
                      "title": "The library that stays open all night",
                      "updated_at": "2026-09-08T22:41:07.220Z",
                      "user": {
                        "created_at": "2026-08-21T09:14:02.511Z",
                        "id": 7,
                        "username": "marenholt"
                      },
                      "user_id": 7,
                      "voted": false,
                      "votes": 4
                    }
                  ],
                  "page": 1,
                  "page_size": 10,
                  "pages": 1,
                  "total": 6
                },
                "schema": {
                  "$ref": "#/components/schemas/PostPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          },
          "422": {
            "description": "The body or the query string didn't validate. FastAPI's own shape: `detail` is a list of `{loc, msg, type}`, one per field."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "What you saved",
        "tags": [
          "Shelf"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A local development server",
      "url": "http://localhost:8000"
    }
  ],
...
}